fix: 修复 TUI 专利撰写/OA答复全链路断链
修复了从 TUI 自然语言对话到专利撰写/OA答复的 6 个断链：

- PatentAgentConfig 补充 injectRulesTools，注册 parse_office_action/
  validate_amendment/analyze_slop 等 6 个 OA 核心工具
- 新增 OrchestrationExtension，在 Handoff 子 Agent Init 阶段注册
  run_orchestration 工具，打通自然语言→编排链路径
- NewOAResponseTool 注入 WithOAProvider，激活 llm_enhance 节点
- /oa 斜杠命令改用 BuildOAResponseGraphWithOpts 并注入 provider
- EvidenceDomainExtension 全局注册并接入 Patent/Legal Agent 配置
- BuildProjectAgent 同步补充 injectRulesTools 和 OrchestrationExtension

所有变更通过 make all（vet+build+test）验证

+++ tui/component/footer.go
@@ -171,7 +171,7 @@ func (f *Footer) Render(width int64) []string {
 	line := strings.Join(parts, " · ")
 
 	// If the rendered line exceeds width, trim from the right.
-	if core.VisibleWidth(line) > int64(width) {
+	if core.VisibleWidth(line) > width {
 		// Fall back to minimal: just show [?] help
 		if len(visible) > 1 {
 			item := visible[0].Items[0]
